# contracts/openapi.json
{
  "components": {
    "schemas": {
      "AiHistoryItem": {
        "properties": {
          "role": {
            "title": "Role",
            "type": "string"
          },
          "text": {
            "maxLength": 4000,
            "title": "Text",
            "type": "string"
          }
        },
        "required": [
          "role",
          "text"
        ],
        "title": "AiHistoryItem",
        "type": "object"
      },
      "AiRequest": {
        "properties": {
          "action": {
            "title": "Action",
            "type": "string"
          },
          "parameters": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/AiRequestParameters"
              },
              {
                "type": "null"
              }
            ]
          }
        },
        "required": [
          "action"
        ],
        "title": "AiRequest",
        "type": "object"
      },
      "AiRequestParameters": {
        "properties": {
          "history": {
            "anyOf": [
              {
                "items": {
                  "$ref": "#/components/schemas/AiHistoryItem"
                },
                "maxItems": 20,
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "History"
          },
          "members": {
            "anyOf": [
              {
                "items": {
                  "$ref": "#/components/schemas/FamilyMember"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Members"
          },
          "periodFrom": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Periodfrom"
          },
          "periodTo": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Periodto"
          },
          "question": {
            "anyOf": [
              {
                "maxLength": 4000,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Question"
          }
        },
        "title": "AiRequestParameters",
        "type": "object"
      },
      "AiResult": {
        "properties": {
          "action": {
            "title": "Action",
            "type": "string"
          },
          "createdAt": {
            "title": "Createdat",
            "type": "string"
          },
          "id": {
            "title": "Id",
            "type": "string"
          },
          "sections": {
            "items": {
              "$ref": "#/components/schemas/AiSection"
            },
            "title": "Sections",
            "type": "array"
          }
        },
        "required": [
          "id",
          "action",
          "createdAt",
          "sections"
        ],
        "title": "AiResult",
        "type": "object"
      },
      "AiSection": {
        "description": "Плоский формат секции — совместим с клиентом.\n\nПоля опциональны, используется только то, что соответствует type:\n- text: text\n- score: value, max\n- list: items\n- products: products\n- chart: labels, values, kind",
        "properties": {
          "items": {
            "anyOf": [
              {
                "items": {
                  "type": "string"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Items"
          },
          "kind": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Kind"
          },
          "labels": {
            "anyOf": [
              {
                "items": {
                  "type": "string"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Labels"
          },
          "max": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Max"
          },
          "products": {
            "anyOf": [
              {
                "items": {
                  "additionalProperties": true,
                  "type": "object"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Products"
          },
          "text": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Text"
          },
          "title": {
            "default": "Ответ",
            "title": "Title",
            "type": "string"
          },
          "type": {
            "title": "Type",
            "type": "string"
          },
          "value": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Value"
          },
          "values": {
            "anyOf": [
              {
                "items": {
                  "type": "number"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Values"
          }
        },
        "required": [
          "type"
        ],
        "title": "AiSection",
        "type": "object"
      },
...
      "CreditsInfo": {
        "properties": {
          "monthly_limit": {
            "title": "Monthly Limit",
            "type": "number"
          },
          "period": {
            "title": "Period",
            "type": "string"
          },
          "period_limit": {
            "title": "Period Limit",
            "type": "number"
          },
          "remaining": {
            "title": "Remaining",
            "type": "number"
          },
          "subscription": {
            "title": "Subscription",
            "type": "boolean"
          },
          "weekly_limit": {
            "title": "Weekly Limit",
            "type": "number"
          }
        },
        "required": [
          "remaining",
          "period_limit",
          "weekly_limit",
          "monthly_limit",
          "period",
          "subscription"
        ],
        "title": "CreditsInfo",
        "type": "object"
      },
...
      "FamilyMember": {
        "properties": {
          "additional_info": {
            "title": "Additional Info",
            "type": "string"
          },
          "age": {
            "title": "Age",
            "type": "integer"
          },
          "gender": {
            "title": "Gender",
            "type": "string"
          },
          "height": {
            "title": "Height",
            "type": "integer"
          },
          "name": {
            "title": "Name",
            "type": "string"
          },
          "weight": {
            "title": "Weight",
            "type": "integer"
          }
        },
        "required": [
          "name",
          "age",
          "height",
          "weight",
          "gender",
          "additional_info"
        ],
        "title": "FamilyMember",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      },
...
    },
    "securitySchemes": {
      "HTTPBearer": {
        "scheme": "bearer",
        "type": "http"
      }
    }
  },
  "info": {
    "title": "Food Spend Tracker API",
    "version": "1.5.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/ai/credits": {
      "get": {
        "operationId": "get_credits_api_ai_credits_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CreditsInfo"
                }
              }
            },
            "description": "Successful Response"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Get Credits",
        "tags": [
          "AI"
        ]
      }
    },
    "/api/ai/history": {
      "get": {
        "operationId": "get_history_api_ai_history_get",
        "parameters": [
          {
            "in": "query",
            "name": "skip",
            "required": false,
            "schema": {
              "default": 0,
              "minimum": 0,
              "title": "Skip",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 30,
              "maximum": 100,
              "minimum": 1,
              "title": "Limit",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/AiResult"
                  },
                  "title": "Response Get History Api Ai History Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Get History",
        "tags": [
          "AI"
        ]
      }
    },
    "/api/ai/history/{report_id}": {
      "delete": {
        "operationId": "delete_report_api_ai_history__report_id__delete",
        "parameters": [
          {
            "in": "path",
            "name": "report_id",
            "required": true,
            "schema": {
              "title": "Report Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Delete Report",
        "tags": [
          "AI"
        ]
      },
      "get": {
        "operationId": "get_report_api_ai_history__report_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "report_id",
            "required": true,
            "schema": {
              "title": "Report Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AiResult"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Get Report",
        "tags": [
          "AI"
        ]
      }
    },
    "/api/ai/run": {
      "post": {
        "operationId": "run_ai_api_ai_run_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AiRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AiResult"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Run Ai",
        "tags": [
          "AI"
        ]
      }
    },
...
  }
}